issue/1029 Renamed classes
