+++ js/adapt-contrib-narrative.js
@@ -1,6 +1,6 @@
 import Adapt from 'core/js/adapt';
-import NarrativeModel from './narrativeModel';
-import NarrativeView from './narrativeView';
+import NarrativeModel from './NarrativeModel';
+import NarrativeView from './NarrativeView';
 
 export default Adapt.register('narrative', {
   model: NarrativeModel,
